+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Changed
+- Add correlation id for all requests sent with the graphql client.
 
 ## [0.3.1] - 2020-11-09